Some further fenagling of the memory layout and vector table
cortex-m-rt's efforts are unfortunately insufficient for this target
because the boot ROM expects a very particular layout including some
special extra values crammed into reserved fields in the vector table.
For now we have some custom code and a custom linker script in a separate
scratch program to start dealing with that; we'll factor it out into a
separate library later.

This still doesn't quite work, for a reason not yet determined. Will
finagle further in later commits.