mimxrt500-rt: cortex-m-rt replacement for mimxrt5xxs
This family has some unusual requirements such as interlacing special
extra values into reserved entries in the vector table, and laying out
flash memory with room for a special configuration block.

To handle that, this is essentially a fork of cortex-m-rt but specialized
for these chips, generating the expected entries in the vector table and
cooperating with code in a board support package to produce a suitable
flash configuration block.