scratch: Blinky blinky
The RTT setup here doesn't seem to be working quite right yet, so instead
we'll use a blinking LED on the eval board as our proof of being able to
run code on this board.

This intentionally uses very simple direct register writes just to make
it easier to inspect in a debugger and verify that it's running correctly.