bootstub: A new approach to startup code
Previously we essentially forked cortex-m-rt just to slightly adapt its
linker script and vector table to match what the RT500 boot rom expects.

That's a really heavy and annoying solution though, so here we're trying
something else: a tiny little bootloader that we can write once at the
start of the flash memory that then jumps into a cortex-m-rt-shaped program
flashed to a later page of flash memory.

With a suitable application linker script that avoids stomping over this
little boot stub (not yet written) we should be able to support
applications linked with cortex-m-rt in the standard way, at the expense
of wasting a bit of flash memory on a redundant vector table and a little
code reset VTOR and jump into the real application.