bootstub: Macro for generating a "flash control block"
The on-chip boot ROM expects to find a block of data at a fixed offset into
the NOR flash to declare that the flash is intended to be used as boot
media.

This new macro "fcb!" generates a static symbol with a special link
section so that a suitably-equipped linker script can place it at the
expected location, which is offset 0x0400 into the flash.