bootstub: two more bootstub macros
bootstub_builtin!() takes no arguments and tries to automatically find the
main application vector table using the __STACK_START symbol that is
conventionally declared when linking with the cortex-m crate.

bootstub_builtin_custom!(path) is similar to bootstub_builtin! but allows
the caller to specify a global symbol to use as the application vector
table. This might be useful for applications that don't use the cortex-m
crate and thus define their own vector table elsewhere.