Replace -rt package with bootstub shimming
Previously we had what was essentially a fork of cortex-m-rt to deal with
this target's unusual requirements for flash layout, including the need
to interleave special values into the vector table. That's annoying though,
because it makes it harder to use idiomatic embedded Rust patterns on
this target.

To avoid that, we'll now use a "boot stub", which effectively reserves the
first 64KiB of flash for some mechanically-generated data and coded that
should satisfy the on-chip boot ROM enough to give us control, and then
we use a little trampoline to start a normal cortex-m-rt-styled program
whose vector table is in a different location.

This solution is not wholly satisfying because it does waste 64KiB of
flash, but it seems like a better tradeoff than entirely forking the
cortex-m-rt package.